paragraph search: improve deduplication of results
