+++ data/heuristics.txt
@@ -2,6 +2,8 @@ incoming
 tagged
 edited
 updated
+modified
+most recently modified
 last update
 last edit
 ©
