+++ lieu.toml
@@ -1,10 +1,10 @@
 [general]
-name = "Forum Search"
-tagline = "search all of forum.merveilles.town"
-placeholder = "Forum Search"
+name = "Merveilles Webring"
 # used by the precrawl command and linked to in /about route
-url = "https://example.com/"
-port = 10099
+url = "https://webring.xxiivv.com/"
+port = 10003
+tagline = "the search for the new—endless"
+placeholder = "Search"
 
 [theme]
 # colors specified in hex (or valid css names) which determine the theme of the lieu instance
@@ -14,7 +14,7 @@ links = "#FF8000" # "#FFCCBB"
 
 [data]
 # the source file should contain the crawl command's output 
-source = "data/source.txt"
+source = "data/crawled.txt"
 # location & name of the sqlite database
 database = "data/searchengine.db"
 # contains words and phrases disqualifying scraped paragraphs from being presented in search results
@@ -33,5 +33,3 @@ bannedSuffixes = "data/banned-suffixes.txt"
 boringWords = "data/boring-words.txt"
 # domains that won't be output as outgoing links
 boringDomains = "data/boring-domains.txt"
-# queries to search for finding preview text
-previewQueryList = "data/preview-query-list.txt"
